third_party/blink/web_tests/storage/indexeddb/transaction-storeNames-required-expected.txt

Test IndexedDB: transaction storeNames required arguments

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "transaction-storeNames-required.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
PASS db.transaction(); threw exception TypeError: Failed to execute 'transaction' on 'IDBDatabase': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
